migrate resources to core mod
