--- mods/core/body_templates.yaml
@@ -1,71 +1,0 @@
-human:
-  torso:
-    chest: 
-      bone_groups:
-      - ribs
-      # TODO: and also
-      organs:
-      - heart
-      - left_lung
-      - right_lung
-      # TODO: and also
-    abdomen:
-      bone_groups:
-        - huy_kammenuy
-        # TODO: idk lol
-      organs:
-      - stomach
-      - liver
-      - intestine
-      # TODO: and also 
-    pelvic:
-      bone_groups:
-      - pelvic
-  head:
-    head:
-      bone_groups:
-      - skull
-      - jaw
-      organs:
-      - brain
-      - left_eye
-      - right_eye
-      # TODO: and also
-  left_arm: &arm
-    arm:
-      bone_groups:
-      - arm
-    forearm:
-      bone_groups:
-      - forearm
-    hand:
-      bone_groups:
-      - hand
-      properties:
-        size: 5
-      flags:
-      - holder
-  right_arm:
-    *arm
-  left_leg: &leg
-    thigh:
-      bone_groups:
-        - thigh
-    knee:
-      bone_groups:
-        - knee
-    leg:
-      bone_groups:
-        - leg
-    foot:
-      bone_groups:
-        - foot
-  right_leg:
-    *leg
-nettle:
-  leaf:
-    leaf:
-      bone_groups:
-        - leaf
-      organs:
-        - leaf
