--- mods/core/entity_templates.yaml
@@ -1,17 +1,0 @@
-nettle:
-  body: nettle
-  sprite: nettle
-  # position: [10, 10, 0]
-  behaviors:
-  - mob
-  - pathfinder
-human:
-  body: human
-  sprite: person
-  sight: 40
-  # position: [1, 1, 0]
-  behaviors:
-  - mob
-  - map_memory
-  - inventory
-  - log
