+++ mods/core/gfx/assets.yaml
