+++ mods/core/gfx/mobs.png
