+++ mods/core/templates/body_templates.yaml
