+++ mods/core/templates/entity_templates.yaml
