+++ mods/core/templates/item_templates.yaml
