+++ src/resources/templates/mod.rs
@@ -1,0 +1,3 @@
+pub mod body;
+pub mod entity;
+pub mod item;
